critical fix for gcc
gcc has a bind type for symbols called "unique" (by readelf -Ws) or
"STB_GNU_UNIQUE" in the gcc documentation. If at least one symbol of
this type is part of a shared object, a dlclose no longer actually
unloads. This behavior can be disabled (cppjit does it for the gcc
builder) with the option "-fno-gnu-unique".

Some details: https://gcc.gnu.org/onlinedocs/gcc/Code-Gen-Options.html

This is highly annoying since the is no proper error and still the
wrong kernel is used. Idea:
- always add the "-fno-gnu-unique"-option to the end of the compile
and link flags for gcc builder? Currently setting the flags requires a
manual addition of this flag (with a hard to fix error if forgotten).